use engine functions in train val scripts
